[feat] #133 댓글 삭제 기능 구현

+++ app/src/main/java/com/android/go/sopt/winey/data/repository/FeedRepositoryImpl.kt
@@ -68,7 +68,7 @@ class FeedRepositoryImpl @Inject constructor(
             feedDataSource.postComment(feedId, requestPostCommentDto).data?.toComment()
         }
 
-    override suspend fun deleteComment(commentId: Int): Result<Unit> =
+    override suspend fun deleteComment(commentId: Long): Result<Unit> =
         runCatching {
             feedDataSource.deleteComment(commentId)
         }
