+++ app/src/main/java/com/android/go/sopt/winey/data/source/FeedDataSource.kt
@@ -41,6 +41,6 @@ class FeedDataSource @Inject constructor(
     ): BaseResponse<ResponsePostCommentDto> =
         feedService.postComment(feedId, requestPostCommentDto)
 
-    suspend fun deleteComment(commentId: Int): BaseResponse<Unit> =
+    suspend fun deleteComment(commentId: Long): BaseResponse<Unit> =
         feedService.deleteComment(commentId)
 }
